Alterar tipo de Id e ajustar cronograma de trabalho
Mudanças significativas incluem:
- O tipo da propriedade `Id` na classe `IbgeSubDistrictDto` foi alterado de `int` para `long`, permitindo um maior intervalo de valores.
- A expressão de cronograma na configuração do trabalho foi modificada para executar no primeiro dia de cada mês às 1:00 AM.
- Na classe `IbgeService`, o valor de `Id` é convertido de `long` para `int`, o que pode causar perda de dados se o valor exceder o limite de um `int`.

+++ Infraestructure/ExternalServices/HttpClients/IbgeService.cs
@@ -232,7 +232,7 @@ public class IbgeService : IIbgeService
             foreach (IbgeSubDistrictDto item in response)
             {
                 var subDistrict = new SubDistricts(
-                    item.Id, 
+                    (int)item.Id, 
                     item.Nome, 
                     districtId);
